tradesense/api/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
libsql-client==0.3.0
python-dotenv==1.0.0
yfinance==0.2.35
beautifulsoup4==4.12.2
requests==2.31.0
werkzeug==3.0.1
